Internal #861: Refactor CumeDist Function
Move the WindowCumeDistExecutor classes to separate files.
